Add messages for users to confirm their action

+++ dogblog/views.py
@@ -1,16 +1,19 @@
-from django.shortcuts import render, get_object_or_404, reverse
+from django.shortcuts import render, get_object_or_404, reverse, redirect
+from django.contrib import messages
 from django.views import generic, View
 from django.http import HttpResponse
 from django.http import HttpResponseRedirect
 from .models import Dog, Competition, DogPhoto, LikePhoto
 from .forms import DogPhotoForm, DogForm
 
-# Create your views here.
+
 def home(request):
     return render(request, 'home.html')
 
+
 def photos(request):
     return render(request, 'photos.html')
+
 
 def add_photo(request):
     if request.method == 'POST':
@@ -45,12 +48,17 @@ def competition(request):
 
 def edit_photo(request, photo_id):
     photo = get_object_or_404(DogPhoto, id=photo_id)
+    if photo.dog.owner != request.user:
+        messages.error(request, 'Not Authorised To Perfom This Action!')
+        return redirect(reverse('competition'))
     if request.method == 'POST':
         form = DogPhotoForm(request.POST, request.FILES, instance=photo)
+        form.fields['dog'].queryset = Dog.objects.filter(owner=request.user)
         if form.is_valid():
             form.save()
             return HttpResponseRedirect(reverse('competition'))
     form = DogPhotoForm(instance=photo)
+    form.fields['dog'].queryset = Dog.objects.filter(owner=request.user)
     context = {
         'form': form
     }
@@ -59,6 +67,9 @@ def edit_photo(request, photo_id):
 
 def delete_photo(request, photo_id):
     photo = get_object_or_404(DogPhoto, id=photo_id)
+    if photo.dog.owner != request.user:
+        messages.error(request, 'Not Authorised To Perfom This Action!')
+        return redirect(reverse('competition'))
     photo.delete()
     return HttpResponseRedirect(reverse('competition'))
 
@@ -69,8 +80,10 @@ class LikeDogPhoto(View):
 
         if photo.likes.filter(id=request.user.id).exists():
             photo.likes.remove(request.user)
+            messages.info(self.request, f"You no longer like {photo.dog.name}'s photo")
         else:
             photo.likes.add(request.user)
+            messages.info(self.request, f"You liked {photo.dog.name}'s photo")
         
         return HttpResponseRedirect(reverse('competition'))
 
@@ -90,7 +103,3 @@ def add_dog(request):
     }
 
     return render(request, 'add_dog.html', context)
-
-
-
-
